refactor: split scalar vectors into separate files
